requirements.txt
asgiref==3.8.1
black==24.3.0
click==8.1.7
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.2.0
Django==5.0.3
django-filter==24.2
django-import-export==3.2.0
django-jazzmin==2.6.1
djangorestframework==3.15.1
et-xmlfile==1.1.0
iniconfig==2.0.0
Markdown==3.6
mypy-extensions==1.0.0
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
psycopg2==2.9.9
pytest==8.1.1
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.6.0
typing_extensions==4.12.0
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
